Raised error if .delete() is given no arguments to prevent accidental deletion

+++ mongodb_dataset/database.py
@@ -77,13 +77,16 @@ class Table:
         """
         Deletes everything that matches
         """
+        if not filter_expr:
+            raise ValueError("Error! Empty filter expression! Call db.clear() if you want to delete everything")
+
         return self.table.delete_many(self._eval_filter_expr(filter_expr))
 
     def clear(self) -> pymongo.results.DeleteResult:
         """
         Clears the entire table
         """
-        return self.delete()
+        return self.table.delete_many({})
 
     def count(self, **filter_expr) -> int:
         """
